Implements recursive command loading for now organized commands
