+++ package.json
@@ -7,6 +7,7 @@
     "eslint": "^8.44.0",
     "pg": "^8.11.1",
     "random-org": "^2.2.0",
+    "recursive-readdir": "^2.2.3",
     "systeminformation": "^5.18.6",
     "typescript": "^5.1.6",
     "winston": "^3.10.0",
@@ -16,6 +17,7 @@
   "devDependencies": {
     "@types/node": "^20.4.1",
     "@types/pg": "^8.10.2",
+    "@types/recursive-readdir": "^2.2.1",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^5.0.0",
