+++ src/commands/DSA/dsaadd.ts
@@ -1,17 +1,15 @@
 import { Message } from "discord.js";
 import type { ILanguage as lang } from "src/types";
-import { Bot } from "../bot.js";
-import { Command } from "../modules/command.js";
+import { Bot } from "../../bot.js";
+import { Command } from "../../modules/command.js";
 
 export default class Dsaadd extends Command {
-  constructor(client: Bot) {
-    super(client);
+  constructor(client: Bot, category: string, name: string) {
+    super(client, category, name);
   }
   help = {
     show: true,
-    name: "dsaadd",
     usage: `${this.prefix}dsaadd <character> [avatar - if it doesn't start with \`http\`, it will be ignored.] <displayed name>`,
-    category: "DSA",
   };
   async run(client: Bot, message: Message, args: string[], language: lang) {
     if (!args || args.length <= 3) {
