+++ src/commands/DSA/new.ts
@@ -7,18 +7,16 @@ import {
   TextChannel,
 } from "discord.js";
 import type { ILanguage as lang } from "src/types";
-import { Bot } from "../bot.js";
-import { Command } from "../modules/command.js";
+import { Bot } from "../../bot.js";
+import { Command } from "../../modules/command.js";
 
 export default class New extends Command {
-  constructor(client: Bot) {
-    super(client);
+  constructor(client: Bot, category: string, name: string) {
+    super(client, category, name);
   }
   help = {
     show: true,
-    name: "new",
     usage: `${this.prefix}new`,
-    category: "DSA",
   };
   run(client: Bot, message: Message, _args: string[], language: lang) {
     let i = 0;
