+++ src/commands/DSA/roll.ts
@@ -1,17 +1,15 @@
 import { EmbedBuilder, Message } from "discord.js";
 import type { ILanguage as lang } from "src/types";
-import { Bot } from "../bot.js";
-import { Command } from "../modules/command.js";
+import { Bot } from "../../bot.js";
+import { Command } from "../../modules/command.js";
 
 export default class Roll extends Command {
-  constructor(client: Bot) {
-    super(client);
+  constructor(client: Bot, category: string, name: string) {
+    super(client, category, name);
   }
   help = {
     show: true,
-    name: "roll",
     usage: `${this.prefix}roll [args]`,
-    category: "Utility",
   };
   async run(client: Bot, message: Message, args: string[], language: lang) {
     const msgauthor: string = message.author.username;
