+++ src/commands/Gifs/cuddle.ts
@@ -1,22 +1,20 @@
 import type { ColorResolvable } from "discord.js";
 import { EmbedBuilder, Message } from "discord.js";
 import type { ILanguage as lang } from "src/types";
-import { Bot } from "../bot.js";
-import { GifCommand } from "../modules/command.js";
+import { Bot } from "../../bot.js";
+import { GifCommand } from "../../modules/command.js";
 
-export default class Pat extends GifCommand {
-  constructor(client: Bot) {
-    super(client);
+export default class Cuddle extends GifCommand {
+  constructor(client: Bot, category: string, name: string) {
+    super(client, category, name);
   }
   help = {
     show: true,
-    name: "pat",
-    usage: `${this.prefix}pat [user]`,
-    category: "Gifs",
+    usage: `${this.prefix}cuddle [user]`,
   };
   async run(client: Bot, message: Message, args: string[], language: lang) {
     const gif: string = await client.db.getgif(
-      "pat",
+      "cuddle",
       await client.db.getgiftype(message.author),
     );
     let userA: string = await client.db.getname(message.author);
@@ -31,21 +29,21 @@ export default class Pat extends GifCommand {
       args,
       language,
     );
-    let responseString = "";
+    let responseString: string;
     if (userB == "") {
       responseString = (
-        await client.random.choice(language.command.pat.singleUser)
+        await client.random.choice(language.command.cuddle.singleUser)
       ).replace(/{a}/g, userA);
     } else {
       responseString = (
-        await client.random.choice(language.command.pat.multiUser)
+        await client.random.choice(language.command.cuddle.multiUser)
       )
         .replace(/{a}/g, userA)
         .replace(/{b}/g, userB);
     }
     const embed = new EmbedBuilder()
       .setImage(gif)
-      .setAuthor({ name: "pat" })
+      .setAuthor({ name: "cuddle" })
       .setDescription(responseString)
       .setColor(color);
     message.channel.send({ embeds: [embed] });
