+++ src/commands/Gifs/name.ts
@@ -1,11 +1,11 @@
 import { Message } from "discord.js";
 import type { ILanguage as lang } from "src/types";
-import { Bot } from "../bot.js";
-import { Command } from "../modules/command.js";
+import { Bot } from "../../bot.js";
+import { Command } from "../../modules/command.js";
 
 export default class Name extends Command {
-  constructor(client: Bot) {
-    super(client);
+  constructor(client: Bot, category: string, name: string) {
+    super(client, category, name);
   }
   run(client: Bot, message: Message, args: string[], language: lang) {
     let newname: string;
@@ -21,8 +21,6 @@ export default class Name extends Command {
   }
   help = {
     show: true,
-    name: "name",
     usage: `${this.prefix}name [name]`,
-    category: "Gifs",
   };
 }
